+++ instapy/util.py
@@ -1935,6 +1935,30 @@ def save_account_progress(browser, username, logger):
         logger.exception('message')
 
 
+def get_epoch_time_diff(time_stamp, logger):
+    try:
+        ''' time diff in seconds from input to now'''
+        log_time = datetime.datetime.strptime(time_stamp, '%Y-%m-%d %H:%M')
+
+        former_epoch = (log_time - datetime.datetime(1970, 1, 1)).total_seconds()
+        cur_epoch = (datetime.datetime.now() - datetime.datetime(1970, 1, 1)).total_seconds()
+
+        return cur_epoch - former_epoch
+    except ValueError:
+        logger.error(
+            "Error occurred while reading timestamp value from database")
+        return None
+
+
+def is_follow_me(browser, person=None):
+    # navigate to profile page if not already in it
+    if person:
+        user_link = "https://www.instagram.com/{}/".format(person)
+        web_address_navigator(browser, user_link)
+
+    return getUserData("graphql.user.follows_viewer", browser)
+
+
 def get_users_from_dialog(old_data, dialog):
     """
     Prepared to work specially with the dynamic data load in the 'Likes'